Implement basic splitting and reassembly using SSS.
This commit is based largely on https://github.com/vincent10400094/scion/tree/feature-packet-splitting.
Not every test passes at the moment, but privacyproxy_test.go is passing.
Basic packet splitting and reassembly works if the MTU is large enough, the message short enough, paths are not changing, and only for a fixed number of paths.